Move TestWatcherStart to integration tests
